requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.2.14
langchain-openai==0.1.22
langchain-chroma==0.1.4
langchain-core==0.2.34
azure-storage-blob==12.22.0
openai==1.42.0
PyPDF2==3.0.1
orjson==3.10.7
pydantic==2.8.2
streamlit==1.37.1
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.4.0
fastapi==0.112.1
uvicorn==0.30.6
pip-system-certs
